+++ core-bundle/src/Resources/contao/drivers/DC_Folder.php
@@ -2703,6 +2703,8 @@ class DC_Folder extends \DataContainer implements \listable, \editable
 			}
 		}
 
+		$arrPublicFolders = array();
+
 		// Process files
 		for ($h=0, $c=count($files); $h<$c; $h++)
 		{
@@ -2753,7 +2755,23 @@ class DC_Folder extends \DataContainer implements \listable, \editable
 
 					if (\Config::get('thumbnails') && ($objFile->isSvgImage || $objFile->height <= \Config::get('gdMaxImgHeight') && $objFile->width <= \Config::get('gdMaxImgWidth')))
 					{
-						$thumbnail .= '<br>' . \Image::getHtml(\System::getContainer()->get('contao.image.image_factory')->create(TL_ROOT . '/' . rawurldecode($currentEncoded), array(400, 50, 'box'))->getUrl(TL_ROOT), '', 'style="margin:0 0 2px -19px"');
+						if (!isset($arrPublicFolders[$objFile->dirname]))
+						{
+							$arrPublicFolders[$objFile->dirname] = file_exists($objFile->dirname . '/.public');
+						}
+
+						$blnIsPublic = $arrPublicFolders[$objFile->dirname];
+
+						// Inline the image if the folder is not public and no preview image will be generated (see #636)
+						if (!$blnIsPublic && ($objFile->height !== null && $objFile->height <= 50 || $objFile->width !== null && $objFile->width <= 400))
+						{
+							$thumbnail .= '<br><img src="' . $objFile->dataUri . '" width="' . $objFile->width . '" height="' . $objFile->height . '" alt="" style="margin:0 0 2px -19px">';
+						}
+						else
+						{
+							$thumbnail .= '<br>' . \Image::getHtml(\System::getContainer()->get('contao.image.image_factory')->create(TL_ROOT . '/' . rawurldecode($currentEncoded), array(400, 50, 'box'))->getUrl(TL_ROOT), '', 'style="margin:0 0 2px -19px"');
+						}
+
 						$importantPart = \System::getContainer()->get('contao.image.image_factory')->create(TL_ROOT . '/' . rawurldecode($currentEncoded))->getImportantPart();
 
 						if ($importantPart->getPosition()->getX() > 0 || $importantPart->getPosition()->getY() > 0 || $importantPart->getSize()->getWidth() < $objFile->width || $importantPart->getSize()->getHeight() < $objFile->height)