+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### DEV
 
+ * Inline small images in protected folders in the file manager (see #636).
  * Correctly encode the URL in the DataContainer::switchToEdit() method (see #762).
  * Fix the parent view drag and drop in Firefox (see #666).
  * Correctly display the search results in the extended tree view (see #739).